+++ CHANGELOG.md
@@ -6,4 +6,5 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
+- Add settings-based header selection (`copyCodeWithContext.headers`)
 - Initial release